+++ src/targets/registry.ts
@@ -105,11 +105,19 @@ export class RegistryTarget extends BaseTarget {
    */
   public getRegistryConfig(): RegistryConfig[] {
     const items = Object.entries(BATCH_KEYS).flatMap(([key, type]) =>
-      Object.entries(this.config[key] || {}).map(([canonicalName, conf]) => ({
-        ...(conf as Record<string, unknown>),
-        type,
-        canonicalName,
-      }))
+      Object.entries(this.config[key] || {}).map(([canonicalName, conf]) => {
+        const config = conf as Record<string, unknown>;
+        const result = Object.assign(Object.create(null), config, {
+          type,
+          canonicalName,
+        });
+
+        if (typeof config.onlyIfPresent === 'string') {
+          result.onlyIfPresent = stringToRegexp(config.onlyIfPresent);
+        }
+
+        return result;
+      })
     );
 
     if (items.length === 0 && this.config.type) {
@@ -296,8 +304,7 @@ export class RegistryTarget extends BaseTarget {
 
     if (
       !registryConfig.urlTemplate &&
-      registryConfig.checksums &&
-      registryConfig.checksums.length === 0
+      !(registryConfig.checksums && registryConfig.checksums.length > 0)
     ) {
       this.logger.warn(
         'No URL template or checksums, not adding any file data'